introduce non dkg message crate
